+++ src/utilities/piral-debug-utils/src/elements.d.ts
@@ -1,0 +1,10 @@
+interface PiralExtensionProps {
+  name: string;
+  params: string;
+}
+
+declare module JSX {
+  interface IntrinsicElements {
+    'piral-extension': React.DetailedHTMLProps<PiralExtensionProps, {}>;
+  }
+}