+++ src/utilities/piral-debug-utils/src/state.ts
@@ -18,7 +18,10 @@ export interface PiralDebugState {
     active: boolean;
     path: string;
   };
-  route: string;
+  route: {
+    path: string;
+    state?: any;
+  } | undefined;
 }
 
 const listeners: Array<() => void> = [];